+++ src/mc/api/strategy/BasicStrategy.hpp
@@ -16,7 +16,7 @@ public:
     return;
   }
 
-  std::pair<aid_t, double> next_transition() const override
+  std::pair<aid_t, int> next_transition() const override
   {
     for (auto const& [aid, actor] : actors_to_run_) {
       /* Only consider actors (1) marked as interleaving by the checker and (2) currently enabled in the application */
@@ -24,9 +24,9 @@ public:
         continue;
       }
 
-      return std::make_pair(aid, 1.0);
+      return std::make_pair(aid, 1);
     }
-    return std::make_pair(-1, 0.0);
+    return std::make_pair(-1, 0);
   }
   void execute_next(aid_t aid, RemoteApp& app) override { return; }
